refactor: remove stateful strategy
