+++ src/application/configuration/app.rs
@@ -1,5 +1,4 @@
 use crate::application::configuration::dto::DurationInSeconds;
-use crate::application::service::auth_service::AuthStrategy;
 use crate::domain::crypto::HashingScheme;
 use lazy_regex::Regex;
 use std::collections::HashMap;
@@ -24,7 +23,6 @@ pub struct AppConfigurationBuilder {
     pub port: Option<String>,
     pub host: Option<String>,
     pub log_level: Option<Level>,
-    pub auth_strategy: Option<AuthStrategy>,
 }
 
 impl AppConfigurationBuilder {
@@ -46,7 +44,6 @@ impl AppConfigurationBuilder {
             port: None,
             host: None,
             log_level: None,
-            auth_strategy: None,
         }
     }
 
@@ -127,11 +124,6 @@ impl AppConfigurationBuilder {
         self
     }
 
-    pub fn auth_strategy(&mut self, value: AuthStrategy) -> &mut Self {
-        self.auth_strategy = Some(value);
-        self
-    }
-
     pub fn load_env(&mut self) -> &mut Self {
         self.super_admin_email = env::var(EnvNames::ADMIN_EMAIL).ok();
         self.super_admin_password = env::var(EnvNames::ADMIN_PASSWORD).ok();
@@ -169,9 +161,6 @@ impl AppConfigurationBuilder {
         self.log_level = env::var(EnvNames::LOG_LEVEL)
             .map(|v| Level::from_str(v.as_str()).unwrap())
             .ok();
-        self.auth_strategy = env::var(EnvNames::AUTH_STRATEGY)
-            .map(|v| AuthStrategy::try_from(v).unwrap())
-            .ok();
 
         self
     }
@@ -198,7 +187,6 @@ impl AppConfigurationBuilder {
                 .rp_duration_in_seconds
                 .clone()
                 .unwrap_or(DurationInSeconds(2592000)),
-            auth_strategy: self.auth_strategy.clone().unwrap_or_default(),
         };
 
         let admin_config = AdminConfig {
@@ -260,7 +248,6 @@ pub struct AppConfiguration {
     port: String,
     host: String,
     log_level: Level,
-    auth_strategy: AuthStrategy,
 }
 
 impl AppConfiguration {
@@ -287,7 +274,6 @@ impl AppConfiguration {
             port: server.port,
             host: server.host,
             log_level: server.log_level,
-            auth_strategy: auth.auth_strategy,
         }
     }
 
@@ -355,10 +341,6 @@ impl AppConfiguration {
         self.rp_duration_in_seconds.clone()
     }
 
-    pub fn auth_strategy(&self) -> AuthStrategy {
-        self.auth_strategy.clone()
-    }
-
     pub fn envs(&self) -> HashMap<String, String> {
         let mut envs = HashMap::new();
 
@@ -410,10 +392,6 @@ impl AppConfiguration {
         envs.insert(EnvNames::PORT.to_owned(), self.port.to_owned());
         envs.insert(EnvNames::HOST.to_owned(), self.host.to_owned());
         envs.insert(EnvNames::LOG_LEVEL.to_owned(), self.log_level.to_string());
-        envs.insert(
-            EnvNames::AUTH_STRATEGY.to_owned(),
-            self.auth_strategy.to_string(),
-        );
 
         envs
     }
@@ -437,7 +415,6 @@ impl EnvNames {
     pub const PORT: &'static str = "PORT";
     pub const HOST: &'static str = "HOST";
     pub const LOG_LEVEL: &'static str = "LOG_LEVEL";
-    pub const AUTH_STRATEGY: &'static str = "AUTH_STRATEGY";
 }
 
 pub struct AuthConfig {
@@ -447,7 +424,6 @@ pub struct AuthConfig {
     pub verification_required: bool,
     pub vr_duration_in_seconds: DurationInSeconds,
     pub rp_duration_in_seconds: DurationInSeconds,
-    pub auth_strategy: AuthStrategy,
 }
 
 pub struct AdminConfig {
