+++ src/application/service/mod.rs
@@ -1,3 +1,1 @@
 pub mod auth_service;
-pub mod stateless_auth_service;
-
